move InvertedIndex opaque enum from inverted_index_ffi to inverted_index::opaque
The module is named `opaque` rather than `ffi` to avoid shadowing the
external `ffi` crate that `inverted_index` depends on.

This will allow the rqe_iterators crate to use this type directly
without having to bring the whole inverted_index_ffi crate triggering
linking issues.
