+++ src/redisearch_rs/c_entrypoint/inverted_index_ffi/cbindgen.toml
@@ -4,6 +4,17 @@ autogen_warning = "/* Warning, this file is autogenerated by cbindgen from `src/
 cpp_compat = true
 pragma_once = true
 
+after_includes = """
+
+/**
+ * An opaque inverted index structure. The actual implementation is determined at runtime based on
+ * the index flags provided when creating the index. This allows us to have a single interface for
+ * all index types while still being able to optimize the storage and performance for each index
+ * type.
+ */
+typedef struct InvertedIndex InvertedIndex;
+"""
+
 trailer = """
 // Create a new inverted index object, with the given flag.
 // The out parameter memsize must be not NULL, the total of allocated memory
